csrf protection to lang delete related #609

+++ src/Ojs/JournalBundle/Controller/LangController.php
@@ -9,6 +9,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Ojs\JournalBundle\Entity\Lang;
 use Ojs\JournalBundle\Form\LangType;
 use Ojs\Common\Controller\OjsController as Controller;
+use Symfony\Component\Security\Core\Exception\TokenNotFoundException;
 
 /**
  * Lang controller.
@@ -27,6 +28,8 @@ class LangController extends Controller
         $grid = $this->get('grid')->setSource($source);
 
         $actionColumn = new ActionsColumn("actions", 'actions');
+        ActionHelper::setup($this->get('security.csrf.token_manager'));
+
         $rowAction[] = ActionHelper::showAction('lang_show', 'id');
         $rowAction[] = ActionHelper::editAction('lang_edit', 'id');
         $rowAction[] = ActionHelper::deleteAction('lang_delete', 'id');
@@ -189,14 +192,21 @@ class LangController extends Controller
     }
 
     /**
-     * Deletes a Lang entity.
-     *
+     * @param Request $request
+     * @param $id
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     * @throws TokenNotFoundException
      */
-    public function deleteAction($id)
+    public function deleteAction(Request $request, $id)
     {
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('OjsJournalBundle:Lang')->find($id);
         $this->throw404IfNotFound($entity);
+
+        $csrf = $this->get('security.csrf.token_manager');
+        $token = $csrf->getToken('lang'.$id);
+        if($token!=$request->get('_token'))
+            throw new TokenNotFoundException("Token Not Found!");
         $em->remove($entity);
         $em->flush();
         $this->successFlashBag('successful.remove');
